+++ js/setup.js
@@ -1,0 +1,17 @@
+//необходимое количество сгенерированных обьектов в массиве (по ТЗ):
+const countOfObjects = 25;
+//количество фотографий (по ТЗ):
+const countOfPhotoMin = 1;
+const countOfPhotoMax = 25;
+//количество комментариев под каждым фото (на свое усмотрение):
+const countOfCommentsMin = 1;
+const countOfCommentsMax = 4;
+//количество лайков (по ТЗ):
+const countOfLikesMin = 15;
+const countOfLikesMax = 200;
+////количество аватаров (по ТЗ):
+const countAvatarMin = 1;
+const countAvatarMax = 6;
+
+export {countOfObjects, countOfPhotoMin, countOfPhotoMax, countOfCommentsMin, countOfCommentsMax,
+  countOfLikesMin, countOfLikesMax, countAvatarMin, countAvatarMax};
